Add the ability for signin up the subjects and unsign the subject to the current logged user Extend subject db model for type of the subject

+++ MI.Server.BusinessLogic/Business/SignupBusiness.cs
@@ -65,43 +65,43 @@ namespace MI.Server.BusinessLogic.Business
                 .ToList();
         }
 
-        public async Task CreateSignup(int studentId, int subjectId)
+        public async Task CreateSignup(UserDb student, int subjectId)
         {
-            UserSubjectsDb existingSignup = await _context.UserSubjects
-                .Where(us => !us.IsDeleted)
-                .FirstOrDefaultAsync(us => us.SubjectId == subjectId && us.UserId == studentId);
+            if (student == null)
+                throw new ArgumentNullException();
+
+            var existingSignup = await _context.UserSubjects
+                .FirstOrDefaultAsync(us => us.SubjectId == subjectId && us.UserId == student.Id);
 
             if (existingSignup != null)
             {
-                throw new NotFoundException($"Student with id {studentId} has already signed up for the subject with id {subjectId}.");
+                throw new NotFoundException($"Student with id {student.Id} has already signed up for the subject with id {subjectId}.");
             }
 
-            UserDb student = await _context.Users
-                .Where(s => !s.IsDeleted)
-                .FirstOrDefaultAsync(s => s.Id == studentId);
-
-            SubjectDb subject = await _context.Subjects
+            var subject = await _context.Subjects
                 .Where(s => !s.IsDeleted)
                 .FirstOrDefaultAsync(s => s.Id == subjectId);
 
             _context.UserSubjects.Add(new UserSubjectsDb()
             {
-                User = student ?? throw new NotFoundException($"Student with id {studentId} does not exist in database."),
+                User = student ?? throw new NotFoundException($"Student with id {student.Id} does not exist in database."),
                 Subject = subject ?? throw new NotFoundException($"Subject with id {subjectId} does not exist in database.")
             });
 
             await _context.SaveChangesAsync();
         }
 
-        public async Task DeleteSignup(int studentId, int subjectId)
+        public async Task DeleteSignup(UserDb student, int subjectId)
         {
+            if (student == null)
+                throw new ArgumentNullException();
+
             UserSubjectsDb signup = await _context.UserSubjects
-                .Where(us => !us.IsDeleted)
-                .FirstOrDefaultAsync(us => us.SubjectId == subjectId && us.UserId == studentId);
+                .FirstOrDefaultAsync(us => us.SubjectId == subjectId && us.UserId == student.Id);
 
             if (signup == null)
             {
-                throw new NotFoundException($"A student with id {studentId} hasn't signed up for a subject with id {subjectId}.");
+                throw new NotFoundException($"A student with id {student.Id} hasn't signed up for a subject with id {subjectId}.");
             }
             _context.UserSubjects.Remove(signup);
